+++ build-logic/src/main/kotlin/io/github/kabos/buildlogic/GradleDsl.kt
@@ -1,0 +1,19 @@
+package io.github.kabos.buildlogic
+
+import com.android.build.gradle.LibraryExtension
+import com.android.build.gradle.TestedExtension
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.configure
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+
+fun Project.kotlin(action: KotlinMultiplatformExtension.() -> Unit) {
+    extensions.configure(action)
+}
+
+fun Project.androidLibrary(action: LibraryExtension.() -> Unit) {
+    extensions.configure(action)
+}
+
+fun Project.android(action: TestedExtension.() -> Unit) {
+    extensions.configure(action)
+}
