+++ build-logic/src/main/kotlin/io/github/kabos/buildlogic/KmpPlugin.kt
@@ -1,0 +1,15 @@
+package io.github.kabos.buildlogic
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class KmpPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target) {
+            with(pluginManager) {
+                apply("org.jetbrains.kotlin.multiplatform")
+            }
+        }
+    }
+}
+
