+++ core/model/build.gradle.kts
@@ -1,11 +1,9 @@
-import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 import org.jetbrains.kotlin.gradle.targets.js.dsl.ExperimentalWasmDsl
 import org.jetbrains.kotlin.gradle.targets.js.webpack.KotlinWebpackConfig
 
 plugins {
-    alias(libs.plugins.kotlinMultiplatform)
-    id("buildlogic.android")
+    id("buildlogic.kmp")
+    id("buildlogic.kmp.android")
 }
 
 android.namespace = "io.github.kabos.core.model"
@@ -22,13 +20,6 @@ kotlin {
                     }
                 }
             }
-        }
-    }
-
-    androidTarget {
-        @OptIn(ExperimentalKotlinGradlePluginApi::class)
-        compilerOptions {
-            jvmTarget.set(JvmTarget.JVM_11)
         }
     }
